pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yuzito"
version = "0.1.0"
description = "Raspberry Pi Camera RTMP Stream"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "picamera2",
    "ffmpeg-python",
    "opencv-python",
    "numpy",
    "flask",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[project.scripts]
yuzito = "src.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"